\name{acml.lme-package}
\alias{acml.lme-package}
\alias{acml.lme}
\docType{package}
\title{
Ascertainment Correction Maximum Likelihood for Linear Mixed Effects.
}
\description{

Estimation of outcome dependent sampling of longitudinal response data for scenarios where all outcome data currently exist, but a targeted sub-study is being planned in order to collect additional key exposure information on a limited number of subjects. Stratified sampling based on specific summaries of individual longitudinal trajectories is required, and the code provides an ascertainment corrected maximum likelihood approach for estimation using the resulting biased sample of subjects.
}

\author{
Shawn Garbett, Jonathan Schildcrout, Dexter Duncan

Vanderbilt University

Maintainer: Shawn Garbett <Shawn.Garbett@Vanderbilt.edu>
}

\references{
Schildcrout JS, Garbett SP, Heagerty PJ: Outcome vector dependent sampling with longitudinal continuous response data: stratified sampling based on summary statistics. Biometrics
}

\details{
\tabular{ll}{
Package: \tab acml\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-12-20\cr
License: \tab GPL-2\cr
}
}

\keyword{ ascertainment corrections, maximum likelihood, linear mixed effects, outcome dependent sampling }
\seealso{
\code{\link[lme4:lme4-package]{lme4}}
}


